+++ src/popup/ExtensionTab.tsx
@@ -77,6 +77,13 @@ export default class ExtensionTab extends React.Component<ExtensionTabProps> {
             experimentName="permissions"
             onToggleExperiment={this.props.onToggleExperiment}
           />
+          <ExperimentSwitch
+            title="Hide on unsupported projects"
+            description="We'll hide the unsupported project banner from appearing on JavaScript/ TypeScript repositories."
+            experiments={this.props.extensionState.experiments}
+            experimentName="hideOnUnsupported"
+            onToggleExperiment={this.props.onToggleExperiment}
+          />
         </FormGroup>
       </div>
     );